requirements.txt
requests>=2.25.1
websocket_client>=1.1.0
setuptools>=57.0.0
certifi>=2020.12.5
click==8.0.3